Reorganize perlhack.pod
Following on an IRC conversation, I've attempted to reorganize
perlhack for greater clarity.  I have only cut and paste blocks
of text and amended section titles and levels.  (I have not addressed
any of the numerous factual issues which remain.)

The resulting guide should be clearer for those trying to skim the
table of contents to understand what is covered in perlhack and
whether it is worth an in-depth read.

I see this change as the first step towards future improvements.