#define some symbols for VC6 compatibility
HWND_MESSAGE and WC_NO_BEST_FIT_CHARS are not defined by the VC6
headers unless WINVER is set to 0x0500 or higher, but then the
headers will generate a lot of warnings because Windows 2000
was still in beta when those headers where released.
PROCESSOR_ARCHITCTURE_AMD64 is not defined at all in the
VC6 headers.