Define our own sockaddr_in6 for VC6
The version defined by the VC6 headers is missing the sin6_scope_id
field, used by Socket.xs.  Ideally VC6 should be used with later
versions of the header files from a Windows Platform SDK, but this
patch at least keeps it compilable with plain VC6.  I have not verified
that the IPv6 code will actually works with this configuration.