regex: Multi-char /i shouldnt match single char []
":\N{LATIN SMALL LIGATURE ST}:" !~ /:[_st]:/i

because it is two character, but previously it did match.  The cause of
this is that the code was using swash_fetch() to test if the fold
matches.  But swash_fetch() only checks if the first character matches,
not all characters, so it was falsely returning true.

This is an intermediate commit, with some cleanup of blocks, comments,
and accompanying indentation of regexec.c to follow immediately.