regexec.c: white space changes due to prev. commit
This is essentially a white-space patch.  Commit
dbaae6dafbe2bb8bf9fcd1e83874b9a683e0290f effectively removed two nests
of blocks.  This finishes that removal, outdents the previous innter
blocks, and reformats the comments for the extra space.

One declaration was moved as a result of removing a block.