regcomp.c: Optimize [cC] char class to EXACTF
A two character character class where the two elements are the folds of
each other can be optimized into an EXACTF regnode.  This should not
change the speed of execution appreciably, except that EXACTF regnodes
are candidates for further optimization by combining with adjacent nodes
of the same type.

This commit brings the optimization level up to somewhat better than
when I started mucking around with ANYOF nodes.