[perl #45147] Issue with the exists function
Perl_av_exists tested to see if regdata magic was present,
but did not have any logic to fetch that data in the positive
key case.  Additionally, in the negative key case, if AvFILL
indicated the key existed, it wouldn't return, and would then
fall through to the logic that treated it like a real array.