regcomp: Disallow multi-char folds in lookbehind
The addition of the ANYOFV regnode to treat multi-char folds in a bracketed
character class has exposed a bug, in which those classes have long been able
to be varying length (due to the multi-char fold), but the compiler wasn't
aware of it.  Now it is, and hence won't allow those which have multi-char
folds to be part of a lookbehind pattern, which requires a constant length.

This patch disallows multi-char folds in a lookbehind bracketed character
class.