Porting/release_managers_guide.pod: check the target before launching nukes
Suggest that people run "git status" *before* they go and nuke any
untracked files with "git clean -dxf".

This avoids mistakes where the release manager has accidentally left
an important but untracked file in the checked out directory. We have
built files we know about in in .gitignore, so the "git status" won't
be noisy unless it should be.