In buildtoc, convert all the do_* functions to taking 2 scalars.
Previously they took a list of lines from the file, and all except
do_manifest() immediately joined it back to a single scalar. Instead, slurp
the file in whole, and do_manifest() split it out.

Also, change the code that invokes do_* to accept only a scalar as a return,
now that all functions return a scalar, rather than (potentially) a list to
join.