t/porting/regen.t can test that META.yml is up to date.
As Porting/makemeta uses regen_lib.pl and rename_if_different(), all that is
needed is to pass it --tap on the command line.

t/porting/regen.t now tests that all files mentioned in the release managers
guide are up to date (and tests will fail if they are not), hence there is now
no need for the release manager to explicitly test that generated files are
not stale.