Fix endless loop in Porting/makemeta.
File::Basename::dirname() returns the current device and directory
on VMS when there is no directory information in the path (i.e.,
you ran out of directories), and the current device and directory
are never ".".  Now that we run makemeta in the test suite, the
expectation that you always get "." was causing the test suite to
hang in a CPU-bound loop.