[perl #83018] TEST PATCH: commit 95ea86d5 also fixes sprintf("%.2f")
This adds some more test cases about the bug that commit
95ea86d5 fixes.

The bug was that in the perl 5.12 branch, sprintf("%.2f") returns ""
whereas it should return "0.00" -- but this only applies in the very
special cases, not eg. when the template has extra characters. The
log message and the tests of that commit only talks about "%.0f", but
I think it would be worth to also test the "%.2f".