Move ANYOF folding from regexec to regcomp
This is for security as well as performance.  It allows Unicode properties to
not be matched case sensitively.  As a result the swash inversion hash is
converted from having utf8 keys to numeric, code point, keys.

It also for the first time fixes the bug where /i doesn't work for a code point
not at the end of a range in a bracketed character class has a multi-character
fold