In IPC::Open3's fd.t, correct the code added in 1f563db471aa8a00.
C<q_Pie_> is not the same as C<q _Pie_> - the former is a bareword starting
with q, the latter a quoting operator. This error was hidden by the code added
in 45a1ce9706434aec to make the test "more forgiving of random stderr output".

Correct the description - fd 1 is STDOUT, not STDIN.

Win32 currently spits out a "Use of uninitialized value" warning, which we need
to take into account when checking that no error messages have been seen.