rename PAD_MAX to PERL_PADSEQ_INTRO
and increase its scope to all the perl core rather than just pad.c.

The scope needs to increase because we'll need to use it in op.c shortly,
and the rename is because it's about to lose any significance as a
numerical value, and just become a magic number to be tested for equality.