Silence compile warnings before uni tables built
The recent move of Unicode folding to the compilation phase caused
spurious warnings during the miniperl build phase of Perl itself before
the Unicode tables get built.  Before the tables are built, Perl is
unable to know about the Unicode semantics (it has ASCII/Latin1
hard-coded in), but was still trying to access the tables.  Now, it
checks and if the tables aren't present uses just the hard-coded
ASCII/Latin1 semantics.