Add Version::Requirements as a dual-life module
Version::Requirements version 0.101020 has been added as a dual-life
module.  It provides a standard library to model and manipulates module
prerequisites and version constraints as defined in the CPAN::Meta::Spec.

It is a prerequisite for CPAN::Meta, which will be used by the Perl
CPAN Toolchain to standardize interactions with META and MYMETA files.