perldiag: retitle Ambiguous use of %c{%s%s}
This is not very helpful:

=item Ambiguous use of %c{%s%s} resolved to %c%s%s

especially since it is functionally identical to the previous entry:

=item Ambiguous use of %c{%s} resolved to %c%s

Not only can diagnostics.pm never find it, but it is hard for human
beings to understand what the different is at first glance, too.

So filling in the second and fourth %s’s with the two possible values
slays a twain of avians with one piece of petrified matter.